backend/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
markdown==3.3.3
django-filter==2.4.0
drf-yasg==1.20.0
dj-database-url==0.5.0
psycopg2-binary==2.8.6
gunicorn
django-heroku
pytz
coverage
coverage-badge
social-auth-app-django==5.0.0
python-jose~=3.3.0
python-dotenv~=0.19.2
drf-jwt
pyjwt
cryptography